requirements.txt
flask
python-dotenv
langchain
pymongo
ollama
openai
serpapi
